% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{acknowledge_revision}
\alias{acknowledge_revision}
\title{Send revision received acknowledgement}
\usage{
acknowledge_revision(article)
}
\arguments{
\item{article}{this is the article id}
}
\description{
Send revision received acknowledgement
}
