% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{acknowledge_submission}
\alias{acknowledge_submission}
\title{Send submission acknowledgements}
\usage{
acknowledge_submission(article)
}
\arguments{
\item{article}{this is the article id}
}
\description{
Send submission acknowledgements
}
