% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{acknowledge_submission_text}
\alias{acknowledge_submission_text}
\title{Send submission acknowledgement drafts}
\usage{
acknowledge_submission_text(article)
}
\arguments{
\item{article}{article id}
}
\description{
Send submission acknowledgement drafts
}
