% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.r
\name{acknowledge_submissions}
\alias{acknowledge_submissions}
\title{Send submission acknowledgement drafts}
\usage{
acknowledge_submissions(drafts)
}
\arguments{
\item{drafts}{list of \code{gmail_draft} objects}
}
\description{
Send submission acknowledgement drafts
}
