% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{reject}
\alias{reject}
\alias{reject_format}
\alias{accept}
\alias{withdraw}
\title{Shortcut for updating status of accept, reject, and withdraw.}
\usage{
reject(article, comments = "", date = Sys.Date())

reject_format(article, comments = "", date = Sys.Date())

accept(article, comments = "", date = Sys.Date())

withdraw(article, comments = "", date = Sys.Date())
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}

\item{comments}{Any additional comments}

\item{date}{Date of status update. If omitted defaults to today.}
}
\description{
This set of functions update the Status field in the DESCRIPTION file and
then draft an email for accepting, rejecting, and withdraw the paper.
}
