% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{actionable_articles}
\alias{actionable_articles}
\title{Show articles that require attention with the corresponding action}
\usage{
actionable_articles(editor, invite = 7, review = 30, verbose = FALSE)
}
\arguments{
\item{editor}{string, initial of an editor or an associate editor, defaults to \code{RJ_EDITOR} env var}

\item{invite}{integer, number of days after which an invite is considered overdue}

\item{review}{integer, number of days after which a review is considered overdue. Note that you can extend this by specifying a due date in the comment, e.g.: "2021-03-01 Agreed (until 2021-05-01)" would allow for two months.}

\item{verbose}{logical, if \code{TRUE} it will always list the full reviewer report for each entry}
}
\description{
Show articles that require attention with the corresponding action
}
