\name{active_articles}
\alias{accepted_articles}
\alias{active_articles}
\title{List articles.}
\usage{
active_articles()

accepted_articles()
}
\description{
List all articles in common directories.
}
\details{
\itemize{ \item \code{active_articles}:
\file{Submissions/}, \file{Accepted/} \item
\code{accepted_articles}: \file{Accepted/} }
}

