% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_workload.R
\name{add_ae}
\alias{add_ae}
\title{Add AE to the DESCRIPTION}
\usage{
add_ae(article, name, date = Sys.Date())
}
\arguments{
\item{article}{article id}

\item{name}{a name used to match AE, can be AE initials, name, github handle, or email}

\item{date}{the date for updating status}
}
\description{
Fuzzy match to find the initial of the AE to fill in the article DESCRIPTION.
Checks that AE term has not ended.
The status field is also updated with a new line of add AE.
}
