% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{AEs}
\alias{AEs}
\alias{detect_AE}
\alias{AE}
\alias{is_AE}
\title{Associate editor (AE) functions}
\usage{
AEs()

detect_AE(path = ".", require = FALSE)

AE(path = ".")

is_AE(path = ".")
}
\arguments{
\item{path}{string, path to the git repository to use
for detection.}

\item{require}{logical, if \code{TRUE} then failing to detect
the AE is considered an error}
}
\value{
\code{AEs} returns a data frame with all associate editors

\code{detect_AE} returns \code{NULL} if not found or the
       row from \code{AEs()} corresponding to the AE

\code{AE} returns \code{NULL} if not an AE or a row from
        \code{AEs()}

\code{is_AE} returns \code{TRUE} if the user if an AE or
        the repository is an AE repository, \code{FALSE} otherwise
}
\description{
\code{AEs()} returns a data frame with information on
associate editors

\code{detect_AE} tries to determine associate editor (AE) from the
remote of the repository in \code{path} or from the git config e-mail.
Note that for editors this will fail.

\code{AE} returns the corresponding row from \code{AEs()} if called
by an associate editor or in an AE repository, otherwise
\code{NULL}. It relies on either \code{RJ_EDITOR} environment
variable to contain the name of the editor or detection from the
git repository pointed to by \code{path} (see
\code{\link{detect_AE}}).

\code{is_AE()} is a shorthand for \code{!is.null(AE())}
}
