% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{article_status_plot}
\alias{article_status_plot}
\title{Generates a status plot for articles submitted in the last few years.}
\usage{
article_status_plot(nyears = 4, save = TRUE)
}
\arguments{
\item{nyears}{The number of years considered. Excludes current year.}

\item{save}{Defaults to TRUE. The plot is saved in the rjournal.github.io/resources folder.}
}
\value{
a ggplot, one bar per year (taken from article id)
}
\description{
Generates a status plot for articles submitted in the last few years.
}
\examples{
\dontrun{
article_status_plot()
}
}
