% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author-email.R
\name{author_emails}
\alias{author_emails}
\title{Extract email from leading author of an issue for emailing}
\usage{
author_emails(issue)
}
\arguments{
\item{issue}{the issue number in the `Proofs` folder, i.e. "2021-2"}
}
\description{
If the email of leading author is not available, the first author with
available email will be used.
}
\examples{
\dontrun{
author_emails("2021-2")
}
}
