% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue.R
\name{build_issue}
\alias{build_issue}
\title{Build an issue}
\usage{
build_issue(id)
}
\arguments{
\item{id}{the id of the issue}
}
\description{
Embeds the article bibliographies as bbl, adds the article import
stanzas to the template, and compiles the proof as PDF.
}
