% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{build_latex}
\alias{build_latex}
\title{Build article from LaTeX}
\usage{
build_latex(
  article,
  share_path = normalizePath("../share", mustWork = TRUE),
  clean = TRUE
)
}
\arguments{
\item{article}{The article to build}

\item{share_path}{???}

\item{clean}{Remove generated files}
}
\description{
Build article from LaTeX
}
