% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_packages_available}
\alias{check_packages_available}
\title{Check that packages mentioned in the text are available on CRAN}
\usage{
check_packages_available(path, name)
}
\arguments{
\item{path}{location of the submission being checked}

\item{name}{name of the tex file containing the article text}
}
\description{
Check that packages mentioned in the text are available on CRAN
}
