% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_wrappers}
\alias{check_wrappers}
\title{Check for the two expected RJwrapper files}
\usage{
check_wrappers(submission_files)
}
\arguments{
\item{submission_files}{a vector of the file names in the submission folder

Check for RJwrapper files
Immediate failure?
Return a warning if either "RJwrapper.tex" and/or "RJwrapper.pdf" not found
Remove the two wrapper files that were checked}
}
\description{
Check for the two expected RJwrapper files
}
