% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_workload.R
\name{corr_author}
\alias{corr_author}
\title{Extract corresponding author from an article}
\usage{
corr_author(article)
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}
}
\description{
Extract corresponding author from an article
}
\examples{
# extract from a single article
corr_author("Submissions/2020-114")

# extract corresponding authors from the active articles
all <- active_articles()
purrr::map_dfr(all, corr_author)
}
