% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{decline_reviewer}
\alias{decline_reviewer}
\alias{agree_reviewer}
\alias{abandon_reviewer}
\title{Update whether the invited reviewer agree, decline, or doesn't replay (abandon) to review}
\usage{
decline_reviewer(article, reviewer_id)

agree_reviewer(article, reviewer_id)

abandon_reviewer(article, reviewer_id)
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}

\item{reviewer_id}{Numeric, the index of the intended reviewer in the Reviewer field.
1 for the first reviewer, 2 for the second}
}
\description{
This function updates the Reviewers field in the DESCRIPTION
}
\examples{
\dontrun{
# first reviewer declined
decline_reviewer("2020-114", reviewer_id = 1)

# second reviewer agreed
agree_reviewer("2020-114", reviewer_id = 2)

# third reviewer doesn't reply and deemed abandon
abandon_reviewer("2020-114", reviwer_id = 3)
}
}
