% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{draft_acknowledge_submissions}
\alias{draft_acknowledge_submissions}
\title{Send submission acknowledgement drafts}
\usage{
draft_acknowledge_submissions(drafts)
}
\arguments{
\item{drafts}{list of \code{gmail_draft} objects}
}
\description{
Send submission acknowledgement drafts
}
