% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{draft_proofing}
\alias{draft_proofing}
\title{Generate proofing email for one article}
\usage{
draft_proofing(article)
}
\arguments{
\item{article}{this is the article id}
}
\description{
Generate proofing email for one article
}
