% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{filter_status}
\alias{filter_status}
\title{Find articles with a given status.}
\usage{
filter_status(articles, status)
}
\arguments{
\item{articles}{A vector of articles, as given by accepted_articles()}

\item{status}{The status you are looking for}
}
\description{
Find articles with a given status.
}
