% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{format.status_list}
\alias{format.status_list}
\title{#' @export
length.status_list <- function(x) length(x)}
\usage{
\method{format}{status_list}(x, ...)
}
\description{
#' @export
length.status_list <- function(x) length(x)
}
