% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{future_ids}
\alias{future_ids}
\title{Generate a new id value.}
\usage{
future_ids(ids, n = 1)
}
\arguments{
\item{ids}{XXX}

\item{n}{No. of ids to generate}
}
\description{
Inspects submissions/, accepted/ and rejected to figure out which
id is next in sequence.
}
