% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{get_accepted_articles}
\alias{get_accepted_articles}
\title{Get list of articles in the Accepted folder to be proofed
This can be used with \code{draft_proofing} to construct
emails to authors on the final version.}
\usage{
get_accepted_articles()
}
\description{
Get list of articles in the Accepted folder to be proofed
This can be used with \code{draft_proofing} to construct
emails to authors on the final version.
}
