% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{get_accepted_but_not_online}
\alias{get_accepted_but_not_online}
\title{Get articles to go online}
\usage{
get_accepted_but_not_online()
}
\description{
Find the articles that are accepted but have not yet been
published to online
}
