% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{get_article_keywords}
\alias{get_article_keywords}
\title{Extract keywords from a submitted article}
\usage{
get_article_keywords(id)
}
\arguments{
\item{id}{the article id}
}
\value{
a vector of the article keywords
}
\description{
Extract keywords from a submitted article
}
