% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_articles_path}
\alias{get_articles_path}
\title{Get the directory of the articles repository, either as set by
set_articles_path() or using git to determine repository root}
\usage{
get_articles_path()
}
\description{
This path is used to point to articles on your file system.
}
