% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{get_md_from_pdf}
\alias{get_md_from_pdf}
\title{Get metadata from a PDF file}
\usage{
get_md_from_pdf(from, final = FALSE)
}
\arguments{
\item{from}{pdf file}

\item{final}{If in a final state, then also include page information.}
}
\description{
Get metadata from a PDF file
}
