% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late.R
\name{get_reviewers}
\alias{get_reviewers}
\title{Find all reviewers for submissions being handled by a given editor}
\usage{
get_reviewers(editor)
}
\arguments{
\item{editor}{The editor or associate editor handling the submissions}
}
\value{
A data frame of reviewers
}
\description{
Find all reviewers for submissions being handled by a given editor
}
