% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{initial_checks}
\alias{initial_checks}
\title{Complete initial checks for submissions.}
\usage{
initial_checks(print_results = TRUE)
}
\arguments{
\item{print_results}{Indicates whether results should be printed to console}
}
\description{
There are several initial checks undertaken for each submission.
These checks reflect the requests made in the author submission guide.
}
\details{
Results will be show in the initial_checks.log file in the submission folder

Set the directory of the articles repository using `set_articles_path`.
}
