% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late.R
\name{late_aes}
\alias{late_aes}
\title{Find late AEs for submissions being handled by a given editor}
\usage{
late_aes(editor)
}
\arguments{
\item{editor}{The editor handling the submissions}
}
\value{
A data frame of AEs who are late in handling a paper.
}
\description{
This should be run regularly and AEs chased up if they are late.
}
\details{
The number of stars has the following meaning:
\itemize{
  \item 1 star: not responded in more than 12 weeks;
  \item 2 stars: not responded in more than 18 weeks;
  \item 3 stars: not responded in more than 24 weeks;
}
Please chase up late AEs.
}
