% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_reviewers.R
\name{late_reviewers}
\alias{late_reviewers}
\title{Find late reviewers for submissions being handled by a given editor}
\usage{
late_reviewers(editor)
}
\arguments{
\item{editor}{The editor or associate editor handling the submissions}
}
\value{
A data frame of reviewers who have not submitted their review on time.
  1 star: invited more than 6 weeks ago;
  2 stars: invited more than 12 weeks ago;
  3 stars: invited more than 18 weeks ago.
Please chase up late reviewers.
If a reviewer has declined, use \code{\link{decline_reviewer}} to remove them from the list.
If you have decided to abandon a reviewer, use \code{\link{abandon_reviewer}}.
}
\description{
This should be run regularly and reviewers chased up if they are late.
}
