% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{list_reviewers}
\alias{list_reviewers}
\alias{decline_reviewer}
\alias{agree_reviewer}
\title{Review helper function}
\usage{
list_reviewers(article)

decline_reviewer(article, reviewer_id)

agree_reviewer(article, reviewer_id)
}
\arguments{
\item{article}{article id, like \code{"2014-01"}}

\item{reviewer_id}{invite just a single reviewer}
}
\description{
Review helper function
}
