% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{log_error}
\alias{log_error}
\title{Produce a log file entry for an error}
\usage{
log_error(text, ..., file = stdout(), .envir = parent.frame())
}
\arguments{
\item{text}{Description of the error that occurred}

\item{file}{the console output directed to the log, using `stdout`}

\item{`...`}{Additional inputs for text passed to the glue function}
}
\description{
Append a line in the log file that details an error
}
