% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{log_note}
\alias{log_note}
\title{Produce a log file entry for a note}
\usage{
log_note(text, ..., file = stdout(), .envir = parent.frame())
}
\arguments{
\item{text}{Description of the note that occurred}

\item{file}{the console output directed to the log, using `stdout`}

\item{`...`}{Additional inputs for text passed to the glue function}
}
\description{
Append a line in the log file that details a note
}
