% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue.R
\name{make_proof}
\alias{make_proof}
\title{Make a proof of an issue}
\usage{
make_proof(id, exec = FALSE)
}
\arguments{
\item{id}{The id of the issue to proof}

\item{exec}{Set to TRUE to make the proof, the default (FALSE) allows a preview of which articles will be moved where.}
}
\description{
The `make_proof()` function is the first step to creating an issue. It moves
the 'proofed' articles from the `Accepted` folder and news articles from
`News_items/{id}` into `Proofs/{id}`.
}
\details{
After the proof is made with this function, `publish_issue()` can be used to
publish these articles into the `rjournal.github.io` repository.
}
