% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late.R
\name{need_decision}
\alias{need_decision}
\title{Find articles that have at least two completed reviews and need a decision}
\usage{
need_decision(editor)
}
\arguments{
\item{editor}{The editor or associate editor handling the submissions}
}
\value{
A data frame of papers needing more reviewers
}
\description{
Returns all articles with at least 2 completed reviews but no decision
This should be run regularly and decisions made if necessary.
}
