% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late.R
\name{need_reviewers}
\alias{need_reviewers}
\title{Find articles that need reviewers for submissions being handled by a given editor.}
\usage{
need_reviewers(editor)
}
\arguments{
\item{editor}{The editor or associate editor handling the submissions}
}
\value{
A data frame of papers needing more reviewers
}
\description{
Returns all articles with fewer than 2 invited reviewers.
This should be run regularly and new reviewers appointed if necessary.
}
