% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{online_metadata_for_article}
\alias{online_metadata_for_article}
\title{Generate metadata for one article.}
\usage{
online_metadata_for_article(x, final = FALSE)
}
\arguments{
\item{x}{An article}

\item{final}{If in a final state, then also include page information.}
}
\description{
Generate metadata for one article.
}
