% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{get_accepted_articles}
\alias{get_accepted_articles}
\alias{draft_proofing}
\alias{proofing_article}
\alias{proofing_article_text}
\title{Functions for proofing articles}
\usage{
get_accepted_articles()

draft_proofing(article, update = TRUE)

proofing_article(drafts)

proofing_article_text(article)
}
\arguments{
\item{article}{article id}

\item{update}{logical, if \code{TRUE} then the status is updated to "out for proofing"}

\item{drafts}{list of \code{gmail_draft} objects}
}
\description{
Functions for proofing articles
}
\details{
\itemize{
   \item{\code{get_accepted_articles()}: get list of articles in the Accepted folder to be proofed. This can be used with \code{draft_proofing} to construct emails to authors on the final version.}
   \item{\code{draft_proofing()}: generate proofing email for one article}
   \item{\code{proofing_article()}: send proofing article emails}
   \item{\code{proofing_article_text()}: writes the email text into the correspondence folder}
}
}
