% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{proofing_article}
\alias{proofing_article}
\title{Send proofing article emails}
\usage{
proofing_article(drafts)
}
\arguments{
\item{drafts}{list of \code{gmail_draft} objects}
}
\description{
Send proofing article emails
}
