% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{proofing_article_text}
\alias{proofing_article_text}
\title{This function writes the email text into the correspondence folder}
\usage{
proofing_article_text(article)
}
\arguments{
\item{article}{article id}
}
\description{
This function writes the email text into the correspondence folder
}
