\name{publish}
\alias{publish}
\title{Publish an article to online first.}
\usage{
publish(article, home = getwd())
}
\arguments{
  \item{article}{article id}

  \item{home}{Location of the articles directory}
}
\description{
After running this, you'll need to update both the article
and the web repository.
}

