% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish_article}
\alias{publish_article}
\title{Publish an article to online first.}
\usage{
publish_article(
  article,
  volume,
  issue,
  home = get_articles_path(),
  legacy = FALSE
)
}
\arguments{
\item{article}{article id}

\item{home}{Location of the articles directory}

\item{legacy}{(Very) old way of referening the R journal}
}
\description{
After running this, you'll need to update both the article and
the web repository.
}
