% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish_news}
\alias{publish_news}
\title{Publish a news article}
\usage{
publish_news(news, volume, issue, home = get_articles_path())
}
\arguments{
\item{news}{File path to the directory containing the news article to publish}

\item{volume}{The volume of the article's issue}

\item{issue}{The issue number of the article's issue}

\item{home}{Location of the articles directory}
}
\description{
This function will publish a news article to the `rjournal.github.io`
repository.
}
\seealso{
`publish_article`, `publish_issue()`
}
