% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_workload.R
\name{reviewer_summary}
\alias{reviewer_summary}
\title{Tabulate reviewer's Accept/Decline incidence}
\usage{
reviewer_summary(articles)
}
\arguments{
\item{articles}{a tibble summary of articles in the accepted and submissions folder. Output of \code{tabulate_articles()}}
}
\description{
The function examine the DESCRIPTION files of articles tabulated by
\code{tabulate_articles} and count the number of accept and decline
of each reviewer.
}
\examples{
\dontrun{
articles <- tabulate_articles()
reviewer_summary(articles)
}
}
