% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj.R
\docType{package}
\name{rj}
\alias{rj-package}
\alias{rj}
\title{R Journal Package}
\description{
This package provides useful functions for the editors of the R journal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rjournal.github.io/rj}
}

}
\keyword{package}
