% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_articles_path}
\alias{set_articles_path}
\title{Set the directory of the articles repository}
\usage{
set_articles_path(path)
}
\description{
This path is used to locate articles on your file system. If this path is not
specified, the path will default to the current working directory if it is
named "articles".
}
