% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue.R
\name{suffix_labels}
\alias{suffix_labels}
\title{Make article labels and bibliographies unique with a suffix}
\usage{
suffix_labels(
  file,
  suffix,
  macros = c("label", "ref", "pageref", "bibitem", "cite", "citep", "citealt", "citet",
    "citealp")
)
}
\arguments{
\item{file}{The file to convert.}

\item{suffix}{A unique suffix to add to label and bib IDs.}
}
\description{
Make article labels and bibliographies unique with a suffix
}
