% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\docType{data}
\name{symbol}
\alias{symbol}
\alias{symbol_utf8}
\title{Various handy symbols to use in a command line UI}
\format{
An object of class \code{list} of length 4.
}
\usage{
symbol
}
\description{
Various handy symbols to use in a command line UI
}
\keyword{datasets}
