% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{tabulate_articles}
\alias{tabulate_articles}
\title{Tabulate article descriptions}
\usage{
tabulate_articles(dirs = c("Accepted", "Submissions"))
}
\arguments{
\item{dirs}{The directories containing articles to be searched and tabulated.
If TRUE, then all directories will be searched.}
}
\description{
Produce a table describing articles in specific directories.
}
