\docType{data}
\name{valid_status}
\alias{valid_status}
\title{A list of all valid statuses.}
\format{\preformatted{ chr [1:17] "submitted" "acknowledged" "needs reviewers" "needs editor" ...
}}
\usage{
valid_status
}
\description{
A list of all valid statuses.
}
\examples{
valid_status
}
\keyword{datasets}

