% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{acknowledge_submission}
\alias{acknowledge_submission}
\title{Send submission acknowledgements}
\usage{
acknowledge_submission(article, editor)
}
\arguments{
\item{article}{this is the article id}

\item{editor}{optional string, if specified, also sets the \code{Editor:} field to that value and the handling editor will be CCd on the e-mail.}
}
\description{
Send submission acknowledgements
}
