% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{reject}
\alias{reject}
\alias{reject_format}
\alias{accept}
\alias{withdraw}
\title{Accept, reject, or withdraw an article}
\usage{
reject(article, comments = "", date = Sys.Date())

reject_format(article, comments = "", date = Sys.Date())

accept(article, comments = "", date = Sys.Date())

withdraw(article, comments = "", date = Sys.Date())
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}

\item{comments}{Any additional comments}

\item{date}{Date of status update. If omitted defaults to today.}
}
\description{
This set of functions update the Status field in the DESCRIPTION file and
then draft an email to accept, reject, or withdraw the paper. It can be seen
as shortcut to \code{update_status} with relevant status plus automatic email drafting.
}
