% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{active_articles}
\alias{active_articles}
\alias{accepted_articles}
\title{List articles.}
\usage{
active_articles(include = c("Submissions", "Accepted"))

accepted_articles()
}
\arguments{
\item{include}{A character vector of directories to include.}
}
\description{
List all articles in common directories.
}
\details{
\itemize{
 \item \code{active_articles}: \file{Submissions/}, \file{Accepted/}
 \item \code{accepted_articles}: \file{Accepted/}
}
}
