% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{add_reviewer}
\alias{add_reviewer}
\title{Add reviewer to DESCRIPTION}
\usage{
add_reviewer(article, name, email, invite = TRUE)
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}

\item{name}{Full name of the reviewer}

\item{email}{Email address of the reviewer}

\item{invite}{Logical, whether to automatically construct an email to invite the reviewer}
}
\description{
This function adds the reviewer information(name and email) to the Reviewers field in the DESCRIPTION.
}
\examples{
\dontrun{
add_reviewer("2020-114", "Przemyslaw Biecek", "przemyslaw.biecek@gmail.com")
}
}
