% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{add_reviewer}
\alias{add_reviewer}
\title{Invite an reviewer}
\usage{
add_reviewer(article, name, email, invite = TRUE)
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}

\item{name}{Full name of the reviewer}

\item{email}{Email address of the reviewer}

\item{invite}{Logical, whether to automatically construct an email to invite the reviewer}
}
\description{
This function adds the reviewer information(name and email) to the reviewers
field in the DESCRIPTION as well as draft an email to invite teh reviewer.
}
\examples{
\dontrun{
add_reviewer("2019-117", "Quiet Quokka", "qqplot@waspot.com")
}
}
