% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{add_reviewer}
\alias{add_reviewer}
\title{Add review to DESCRIPTION}
\usage{
add_reviewer(article, name, email, invite = TRUE, open_email = TRUE)
}
\arguments{
\item{article}{article id, like \code{"2014-01"}}

\item{name}{display name, optional}

\item{email}{email address}

\item{invite}{Automatically construct email}

\item{open_email}{logical value; should the function attempt to
open the users email program; default is TRUE}
}
\description{
Add a reviewers name to the DESCRIPTION file
}
