% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_workload.R
\name{ae_workload}
\alias{ae_workload}
\title{Check the number of articles an AE is currently working on}
\usage{
ae_workload()
}
\description{
The function Github REST API to query the ae prefixed repository and extract the content in the submissions folder.
}
