% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_workload.R
\name{ae_workload}
\alias{ae_workload}
\alias{get_AE}
\title{Check the number of articles an AE is currently working on}
\usage{
ae_workload()

get_AE(x)
}
\arguments{
\item{x}{An article object}
}
\description{
This will examine the DESCRIPTION files for articles in
the Submissions folder, check articles that have status
"with AE".
}
