% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{get_submissions}
\alias{get_submissions}
\title{Download submissions from gmail.}
\usage{
get_submissions()
}
\value{
list of \code{gmail_draft} objects for
  \code{\link{send_acknowledgements}}
}
\description{
Assumes submissions to be processed are unread emails in the
rjournal.submission@gmail.com inbox.  For authorization to work,
you must be logged into the rjournal.submission account before
calling this function.
}
\section{Process}{

The function does three things automatically:
\enumerate{
 \item Downloads and extracts submissions into appropriate directories.
 \item Marks submissions as "read" in the inbox.
 \item Uploads acknowledgement emails to gmail account as drafts.
}
The user (editor-in-chief) then:
\enumerate{
 \item Ensures that the files have unzipped correctly (some authors
   incorrectly upload .rar or .tar.gz files) and that the latex
   compiles
 \item Manually sends the draft emails
}
}

