% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{invite_reviewers}
\alias{invite_reviewers}
\alias{invite_reviewer}
\alias{add_review}
\title{Invite reviewer(s).}
\usage{
invite_reviewers(article, prefix = "1")

invite_reviewer(article, reviewer_id, prefix = "1")

add_review(article, reviewer_id, review, recommend = NULL)
}
\arguments{
\item{article}{article id, like \code{"2014-01"}}

\item{prefix}{prefix added to start file name - used to distinguish
between multiple rounds of reviewers (if needed)}

\item{reviewer_id}{invite just a single reviewer}

\item{review}{Path to the review, e.g. pdf, txt, or docx format}

\item{recommend}{Summary of review. Accept, minor, major, reject}
}
\description{
Once you have added reviewers to the DESCRIPTION file, you can use
this function to draft invite emails to them all. As well as drafting
the emails, it also caches them locally in the \code{correspodence/}
directory of the corresponding article.
}
