% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late.R
\name{late_reviewers}
\alias{late_reviewers}
\title{Find late reviewers for submissions being handled by a given editor}
\usage{
late_reviewers(editor)
}
\arguments{
\item{editor}{The editor or associate editor handling the submissions}
}
\value{
A data frame of reviewers who have not responded or not submitted their review on time.
}
\description{
This should be run regularly and reviewers chased up if they are late.
}
\details{
The number of stars has the following meaning:
\itemize{
  \item 1 star: not responded in more than a week or accepted more than 4 weeks ago;
  \item 2 stars: not responded in more than 2 weeks or accepted more than 8 weeks ago;
  \item 3 stars: not responded in more than 3 weeks or accepted more than 12 weeks ago;
}
Please chase up late reviewers.
If a reviewer has declined, use \code{\link{decline_reviewer}} to remove them from the list.
If you have decided to abandon a reviewer, use \code{\link{abandon_reviewer}}.
}
