% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_keywords}
\alias{match_keywords}
\title{Match reviewers for a submitted article based on keywords}
\usage{
match_keywords(id, n = 5)
}
\arguments{
\item{id}{the article id in the description file}

\item{n}{numeric; number of reviewer to display}
}
\value{
\itemize{
\item{\code{match_keywords}: a tibble of potential reviewers for the article}
\item{\code{get_article_keywords}: a list with 2 elements: 1) a vector of keywords, and 2) a vector of author names}
\item{\code{get_reviewer_keywords}: a tibble of reviewer information from the googlesheet. Variables include email, gname (given name), fname (family name), and keywords}
}
}
\description{
\code{match_keywords} matches an article to potential reviewers through
  first extract the keywords of the article (\code{get_article_keywords})
  and then query the reviewer googlesheet for registered reviewers and their
  interested reviewing areas (\code{get_reviewer_keywords}). Notice that a
  googlesheet authenticate, with your email address printed, will first pop up
  to verify the access to the reviewer googlesheet.
}
\details{
All the reviewers are ranked based on the number of matching keywords
  and when there is a tie, a random draw is used.

  For example, an article A has 3 keywords. Two reviewers have all the 3 keywords matched,
  5 reviewers have 2 matches, and another 10 have 1 match. To get 5 reviewers for article A,
  both reviewers with 3 matches are in and a random draw, among the five reviewers with 2 matches,
  is used to fill the remaining 3 places.
}
\examples{
\dontrun{
match_keywords("2021-13")
match_keywords("2021-13", n = 10)
}

}
