% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{online_metadata_for_article}
\alias{online_metadata_for_article}
\title{Generate metadata for one article.}
\usage{
online_metadata_for_article(x, final = FALSE)
}
\description{
Generate metadata for one article.
}
