% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publish an article to online first.}
\usage{
publish(article, home = getwd(), legacy = FALSE)
}
\arguments{
\item{article}{article id}

\item{home}{Location of the articles directory}
}
\description{
After running this, you'll need to update both the article and
the web repository.
}
