% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue.R
\name{publish_issue}
\alias{publish_issue}
\title{Publish an issue}
\usage{
publish_issue(
  id,
  web_path = file.path("..", "rjournal.github.io"),
  post_file = c(foundation = 1, cran = 1, bioc = 1, ch = 1)
)
}
\arguments{
\item{id}{the id of the issue}

\item{web_path}{path to the rjournal.github.io checkout}
}
\description{
Generates per-article PDFs and copies them to the website, located
at \code{web_path}. Removes the published articles from the
accepted directory. Generates the necessary metadata and updates
the website configuration.
}
\details{
This depends on the pdftools CRAN package, which in turn depends on
the poppler system library. It also requires the command line
program pdftk (distributed as PDFtk Server).
}
