% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate a status report.}
\usage{
report(articles = active_articles())
}
\arguments{
\item{articles}{list of articles to generate report for. Defaults to
all active reports in \file{Submissions/}.}
}
\description{
This should be run weekly.
}
