% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{list_reviewers}
\alias{list_reviewers}
\alias{reviewer_status}
\title{Summarise reviewers' progression of an article}
\usage{
list_reviewers(article)

reviewer_status(article)
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}
}
\description{
This function summarises the status of reviewers who are willing to review
for a particular article.
}
\examples{
\dontrun{list_reviewers("Submissions/2020-114")}
\dontrun{
reviewer_status("Submissions/2020-114")
}
}
