% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_workload.R
\name{reviewer_summary}
\alias{reviewer_summary}
\title{Summarise the reviewer's agree/decline ratio based on past invites}
\usage{
reviewer_summary(articles)
}
\arguments{
\item{articles}{a tibble summary of articles in the accepted and submissions folder. Output of \code{tabulate_articles()}}
}
\description{
The function pulls out the agree/decline incidence of all the reviewers
based on past invite and calculate the agree percentage of each reviewer.
Use \code{tabulate_articles} first to tabulate all the articles in a particular directory
and then apply this function.
}
\examples{
\dontrun{
articles <- tabulate_articles()
reviewer_summary(articles)
}
}
