% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{summarise_articles}
\alias{summarise_articles}
\alias{get_assignments}
\title{Summarise intray
Summarise editors current in-tray}
\usage{
summarise_articles(editor = NULL, rejected = FALSE)

get_assignments(editor)
}
\arguments{
\item{editor}{Editors initials. If \code{NULL}, looks for the
environment variable \code{RJ_EDITOR}.}

\item{rejected}{Default \code{FALSE}. If \code{TRUE}, show
rejected articles.}
}
\description{
Summarise intray
Summarise editors current in-tray
}
