% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{update_status}
\alias{update_status}
\title{Update the status field in the DESCRIPTION}
\usage{
update_status(article, status, comments = "", date = Sys.Date())
}
\arguments{
\item{article}{Article id, like \code{"2014-01"}}

\item{status}{new status to add, see details section for more}

\item{comments}{Any additional comments}

\item{date}{Date of status update. If omitted defaults to today.}
}
\description{
This function updates the Status field in the DESCRIPTION.
\code{reject}, \code{accept} and \code{withdraw} update the status,
move the file to the correct directory and draft an email from a
template of the corresponding name. Use \code{valid_status} to check
for possible statuses.
}
\details{
Possible status:

\describe{
 \item{AE making recommendation}{"AE: major revision", "AE: minor revision", "AE: accept", and "AE: reject"}
 \item{Editors}{"with AE", "accepted", "rejected", "withdraw"}
}
}
\examples{
\dontrun{
update_status("2020-114", status = "AE: major revision")
}
}
