% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\docType{data}
\name{valid_reviewer_status}
\alias{valid_reviewer_status}
\title{A list of all valid reviewer statuses.}
\format{
An object of class \code{character} of length 9.
}
\usage{
valid_reviewer_status
}
\description{
A list of all valid reviewer statuses.
}
\examples{
valid_reviewer_status
}
\keyword{datasets}
