% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{valid_status}
\alias{valid_status}
\title{A list of all valid statuses.}
\format{
An object of class \code{character} of length 26.
}
\usage{
valid_status
}
\description{
A list of all valid statuses.
}
\examples{
valid_status
}
\keyword{datasets}
